[tracing] Roll perfetto
Roll perfetto to get the fix for a TSAN race.

We need to override a new method on the producer, ClearIncrementalState.
It is only called if the producer opts into it, which we do not so we
add UNREACHABLE() in the implementation.

Cq-Include-Trybots: luci.v8.try:v8_linux64_perfetto_dbg_ng
Bug: v8:8339
Change-Id: I782989f32b511e98479f6637148e3bdd78a6578a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/1615461
Reviewed-by: Jakob Gruber <jgruber@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#61584}