[Android] Don't apply offset to library start address for libs mapped from APK.
On Android, we mmap the .so directly from the APK, therefore the file offset
in this case refers to the offset of the .so in the APK, not the offset in
the .so file itself. As such, moving the start position of the library based
on this gives incorrect results when later symbolizing the addresses based on
the .so.

TBR=ulan@chromium.org

Change-Id: I16f78c91106dc85334a8b0034c96b19442af7b05
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/1627545
Reviewed-by: Ross McIlroy <rmcilroy@chromium.org>
Commit-Queue: Ross McIlroy <rmcilroy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#61803}